Rewrite. Tested with players.
